Extract the migration tool to its own java source
